+++ src/sll/string_table.c
@@ -24,22 +24,6 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_index_t sll_add_string(sll_string_t
 
 
 
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_index_t sll_add_string_object(sll_string_table_t* st,sll_object_t* v){
-	SLL_ASSERT(SLL_OBJECT_GET_TYPE(v)==SLL_OBJECT_TYPE_STRING);
-	for (sll_string_index_t i=0;i<st->l;i++){
-		if (sll_string_equal(st->dt+i,&(v->dt.s))){
-			return i;
-		}
-	}
-	st->l++;
-	st->dt=sll_reallocate(st->dt,st->l*sizeof(sll_string_t));
-	v->t|=OBJECT_EXTERNAL_STRING;
-	*(st->dt+st->l-1)=v->dt.s;
-	return st->l-1;
-}
-
-
-
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_index_t sll_create_string(sll_string_table_t* st,const sll_char_t* dt,sll_string_length_t l){
 	sll_string_t n;
 	sll_string_from_pointer_length(dt,l,&n);